backend/requirements.txt
Flask>=2.3.0
Flask-CORS>=4.0.0
python-dotenv>=1.0.0
Pillow>=10.0.0
imagehash>=4.3.1
pytest>=7.4.0
pytest-cov>=4.1.0
black>=23.9.0
flake8>=6.1.0
